Add the modal to choose the plug
