--- src/components/modal/styles.js
@@ -1,47 +1,0 @@
-import { StyleSheet } from 'react-native';
-
-const styles = StyleSheet.create({
-    centeredView: {
-      flex: 1,
-      justifyContent: "center",
-      alignItems: "center",
-      marginTop: 22
-    },
-    modalView: {
-      margin: 20,
-      backgroundColor: "white",
-      borderRadius: 20,
-      padding: 35,
-      alignItems: "center",
-      shadowColor: "#000",
-      shadowOffset: {
-        width: 0,
-        height: 2
-      },
-      shadowOpacity: 0.25,
-      shadowRadius: 4,
-      elevation: 5,
-    },
-    buttonClose:{
-      backgroundColor: 'black',
-      borderRadius: 10,
-      marginTop: 200,
-      borderWidth: 2,
-      padding: 10,
-      elevation: 2,
-      width: 100
-  },
-    buttonOpen: {
-      backgroundColor: "#59595912"
-    },
-    textStyle: {
-      color: "black",
-      fontWeight: "bold",
-      textAlign: "center"
-    },
-    modalText: {
-      marginBottom: 15,
-      textAlign: "center"
-    }
-  });
-
